+++ Input/InputSpotlightDecorator.js
@@ -128,7 +128,7 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @public
 			 */
 			spotlightDisabled: PropTypes.bool
-		}
+		};
 
 		constructor (props) {
 			super(props);
@@ -188,25 +188,25 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 					this.paused.resume();
 				}
 			}
-		}
+		};
 
 		focus = (focused, node) => {
 			this.setState({focused, node});
-		}
+		};
 
 		blur = () => {
 			this.setState((state) => (
 				state.focused || state.node ? {focused: null, node: null} : null
 			));
-		}
+		};
 
 		focusDecorator = (decorator) => {
 			this.focus('decorator', decorator);
-		}
+		};
 
 		focusInput = (decorator) => {
 			this.focus('input', decorator.querySelector('input'));
-		}
+		};
 
 		onBlur = (ev) => {
 			if (!this.props.autoFocus) {
@@ -233,7 +233,7 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 					this.blur();
 				}
 			}
-		}
+		};
 
 		onMouseDown = (ev) => {
 			const {disabled, spotlightDisabled} = this.props;
@@ -246,7 +246,7 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 
 			forwardMouseDown(ev, this.props);
-		}
+		};
 
 		onFocus = (ev) => {
 			forwardFocus(ev, this.props);
@@ -257,10 +257,10 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				this.focusInput(ev.currentTarget);
 				ev.stopPropagation();
 			}
-		}
+		};
 
 		onKeyDown (ev) {
-			const {currentTarget, keyCode, preventDefault, target} = ev;
+			const {currentTarget, keyCode, target} = ev;
 
 			// cache the target if this is the first keyDown event to ensure the component had focus
 			// when the key interaction started
@@ -286,7 +286,7 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 				// prevent modifying the value via 5-way for numeric fields
 				if ((isUp || isDown) && target.type === 'number') {
-					preventDefault();
+					ev.preventDefault();
 				}
 
 				if (shouldSpotlightMove) {
@@ -317,7 +317,7 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 		onKeyUp = (ev) => {
 			const {dismissOnEnter} = this.props;
-			const {currentTarget, keyCode, preventDefault, target} = ev;
+			const {currentTarget, keyCode, target} = ev;
 
 			// verify that we have a matching pair of key down/up events to avoid adjusting focus
 			// when the component received focus mid-press
@@ -327,14 +327,14 @@ const InputSpotlightDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				if (this.state.focused === 'input' && dismissOnEnter && is('enter', keyCode)) {
 					this.focusDecorator(currentTarget);
 					// prevent Enter onKeyPress which triggers an onMouseDown via Spotlight
-					preventDefault();
+					ev.preventDefault();
 				} else if (this.state.focused !== 'input' && is('enter', keyCode)) {
 					this.focusInput(currentTarget);
 				}
 			}
 
 			forwardKeyUp(ev, this.props);
-		}
+		};
 
 		setDownTarget (ev) {
 			const {repeat, target} = ev;