+++ VideoPlayer/VideoPlayer.js
@@ -154,7 +154,7 @@ const AnnounceState = {
  * @public
  */
 const VideoPlayerBase = class extends React.Component {
-	static displayName = 'VideoPlayerBase'
+	static displayName = 'VideoPlayerBase';
 
 	static propTypes = /** @lends moonstone/VideoPlayer.VideoPlayerBase.prototype */ {
 		/**
@@ -610,7 +610,7 @@ const VideoPlayerBase = class extends React.Component {
 		 * * `pause()` - pause video
 		 * * `load()` - load video
 		 *
-		 * The [`source`]{@link moonstone/VideoPlayer.VideoBase.source} property is passed to
+		 * The [`source`]{@link moonstone/VideoPlayer.Video.source} property is passed to
 		 * the video component as a child node.
 		 *
 		 * @type {Component|Element}
@@ -618,9 +618,9 @@ const VideoPlayerBase = class extends React.Component {
 		 * @public
 		 */
 		videoComponent: EnactPropTypes.componentOverride
-	}
+	};
 
-	static contextType = FloatingLayerContext
+	static contextType = FloatingLayerContext;
 
 	static defaultProps = {
 		autoCloseTimeout: 5000,
@@ -637,7 +637,7 @@ const VideoPlayerBase = class extends React.Component {
 		spotlightId: 'videoPlayer',
 		titleHideDelay: 5000,
 		videoComponent: Media
-	}
+	};
 
 	constructor (props) {
 		super(props);
@@ -816,11 +816,11 @@ const VideoPlayerBase = class extends React.Component {
 	// Internal Methods
 	//
 
-	announceJob = new Job(msg => (this.announceRef && this.announceRef.announce(msg)), 200)
+	announceJob = new Job(msg => (this.announceRef && this.announceRef.announce(msg)), 200);
 
 	announce = (msg) => {
 		this.announceJob.start(msg);
-	}
+	};
 
 	getHeightForElement = (elementName) => {
 		const element = this.player.querySelector(`.${css[elementName]}`);
@@ -829,12 +829,12 @@ const VideoPlayerBase = class extends React.Component {
 		} else {
 			return 0;
 		}
-	}
+	};
 
 	activityDetected = () => {
 		// console.count('activityDetected');
 		this.startAutoCloseTimeout();
-	}
+	};
 
 	startAutoCloseTimeout = () => {
 		// If this.state.more is used as a reference for when this function should fire, timing for
@@ -843,15 +843,15 @@ const VideoPlayerBase = class extends React.Component {
 		if (this.props.autoCloseTimeout && this.state.mediaControlsVisible) {
 			this.autoCloseJob.startAfter(this.props.autoCloseTimeout);
 		}
-	}
+	};
 
 	stopAutoCloseTimeout = () => {
 		this.autoCloseJob.stop();
-	}
+	};
 
 	generateId = () => {
 		return Math.random().toString(36).substr(2, 8);
-	}
+	};
 
 	isTimeBeyondSelection (time) {
 		const {selection} = this.props;
@@ -889,7 +889,7 @@ const VideoPlayerBase = class extends React.Component {
 		}
 
 		return true;
-	}
+	};
 
 	/**
 	 * Shows media controls.
@@ -925,7 +925,7 @@ const VideoPlayerBase = class extends React.Component {
 				titleVisible: true
 			};
 		});
-	}
+	};
 
 	/**
 	 * Hides media controls.
@@ -948,7 +948,7 @@ const VideoPlayerBase = class extends React.Component {
 			infoVisible: false
 		});
 		this.markAnnounceRead();
-	}
+	};
 
 	/**
 	 * Toggles the media controls.
@@ -963,7 +963,7 @@ const VideoPlayerBase = class extends React.Component {
 		} else {
 			this.showControls();
 		}
-	}
+	};
 
 	doAutoClose = () => {
 		this.stopDelayedFeedbackHide();
@@ -975,35 +975,35 @@ const VideoPlayerBase = class extends React.Component {
 			infoVisible: false
 		}));
 		this.markAnnounceRead();
-	}
+	};
 
-	autoCloseJob = new Job(this.doAutoClose)
+	autoCloseJob = new Job(this.doAutoClose);
 
 	startDelayedTitleHide = () => {
 		if (this.props.titleHideDelay) {
 			this.hideTitleJob.startAfter(this.props.titleHideDelay);
 		}
-	}
+	};
 
 	stopDelayedTitleHide = () => {
 		this.hideTitleJob.stop();
-	}
+	};
 
 	hideTitle = () => {
 		this.setState({titleVisible: false});
-	}
+	};
 
-	hideTitleJob = new Job(this.hideTitle)
+	hideTitleJob = new Job(this.hideTitle);
 
 	startDelayedFeedbackHide = () => {
 		if (this.props.feedbackHideDelay) {
 			this.hideFeedbackJob.startAfter(this.props.feedbackHideDelay);
 		}
-	}
+	};
 
 	stopDelayedFeedbackHide = () => {
 		this.hideFeedbackJob.stop();
-	}
+	};
 
 	showFeedback = () => {
 		if (this.state.mediaControlsVisible) {
@@ -1020,7 +1020,7 @@ const VideoPlayerBase = class extends React.Component {
 				}));
 			}
 		}
-	}
+	};
 
 	hideFeedback = () => {
 		if (this.state.feedbackVisible && this.state.feedbackAction !== 'focus') {
@@ -1029,19 +1029,19 @@ const VideoPlayerBase = class extends React.Component {
 				feedbackAction: 'idle'
 			});
 		}
-	}
+	};
 
-	hideFeedbackJob = new Job(this.hideFeedback)
+	hideFeedbackJob = new Job(this.hideFeedback);
 
 	startDelayedMiniFeedbackHide = (delay = this.props.miniFeedbackHideDelay) => {
 		if (delay) {
 			this.hideMiniFeedbackJob.startAfter(delay);
 		}
-	}
+	};
 
 	stopDelayedMiniFeedbackHide = () => {
 		this.hideMiniFeedbackJob.stop();
-	}
+	};
 
 	hideMiniFeedback = () => {
 		if (this.state.miniFeedbackVisible) {
@@ -1051,21 +1051,21 @@ const VideoPlayerBase = class extends React.Component {
 				miniFeedbackVisible: false
 			});
 		}
-	}
+	};
 
-	hideMiniFeedbackJob = new Job(this.hideMiniFeedback)
+	hideMiniFeedbackJob = new Job(this.hideMiniFeedback);
 
-	handle = handle.bind(this)
+	handle = handle.bind(this);
 
 	showControlsFromPointer = () => {
 		Spotlight.setPointerMode(false);
 		this.showControls();
-	}
+	};
 
 	clearPulsedPlayback = () => {
 		this.pulsedPlaybackRate = null;
 		this.pulsedPlaybackState = null;
-	}
+	};
 
 	// only show mini feedback if playback controls are invoked by a key event
 	shouldShowMiniFeedback = (ev) => {
@@ -1073,7 +1073,7 @@ const VideoPlayerBase = class extends React.Component {
 			this.showMiniFeedback = true;
 		}
 		return true;
-	}
+	};
 
 	handleLoadStart = () => {
 		this.firstPlayReadFlag = true;
@@ -1094,31 +1094,31 @@ const VideoPlayerBase = class extends React.Component {
 				this.showControls();
 			}
 		}
-	}
+	};
 
 	handlePlay = this.handle(
 		forwardPlay,
 		this.shouldShowMiniFeedback,
 		() => this.play()
-	)
+	);
 
 	handlePause = this.handle(
 		forwardPause,
 		this.shouldShowMiniFeedback,
 		() => this.pause()
-	)
+	);
 
 	handleRewind = this.handle(
 		forwardRewind,
 		this.shouldShowMiniFeedback,
-		() => this.rewind(),
-	)
+		() => this.rewind()
+	);
 
 	handleFastForward = this.handle(
 		forwardFastForward,
 		this.shouldShowMiniFeedback,
 		() => this.fastForward()
-	)
+	);
 
 	handleJump = ({keyCode}) => {
 		if (this.props.seekDisabled) {
@@ -1133,7 +1133,7 @@ const VideoPlayerBase = class extends React.Component {
 			this.jump(jumpBy);
 			this.announceJob.startAfter(500, secondsToTime(this.video.currentTime, getDurFmt(this.props.locale), {includeHour: true}));
 		}
-	}
+	};
 
 	handleGlobalKeyDown = this.handle(
 		returnsTrue(this.activityDetected),
@@ -1147,7 +1147,7 @@ const VideoPlayerBase = class extends React.Component {
 		preventDefault,
 		stopImmediate,
 		this.showControlsFromPointer
-	)
+	);
 
 	//
 	// Media Interaction Methods
@@ -1181,7 +1181,7 @@ const VideoPlayerBase = class extends React.Component {
 		}
 
 		this.setState(updatedState);
-	}
+	};
 
 	renderBottomControl = new Job(() => {
 		if (!this.state.bottomControlsRendered) {
@@ -1207,7 +1207,7 @@ const VideoPlayerBase = class extends React.Component {
 			proportionLoaded  : this.state.proportionLoaded,
 			proportionPlayed  : this.state.proportionPlayed
 		};
-	}
+	};
 
 	/**
 	 * The primary means of interacting with the `<video>` element.
@@ -1222,7 +1222,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.showFeedback();
 		this.startDelayedFeedbackHide();
 		this.video[action](props);
-	}
+	};
 
 	/**
 	 * Programmatically plays the current media.
@@ -1244,7 +1244,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.send('play');
 		this.announce($L('Play'));
 		this.startDelayedMiniFeedbackHide(5000);
-	}
+	};
 
 	/**
 	 * Programmatically plays the current media.
@@ -1266,7 +1266,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.send('pause');
 		this.announce($L('Pause'));
 		this.stopDelayedMiniFeedbackHide();
-	}
+	};
 
 	/**
 	 * Set the media playback time index
@@ -1282,7 +1282,7 @@ const VideoPlayerBase = class extends React.Component {
 		} else {
 			forward('onSeekFailed', {}, this.props);
 		}
-	}
+	};
 
 	/**
 	 * Step a given amount of time away from the current playback position.
@@ -1304,7 +1304,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.startDelayedFeedbackHide();
 		this.seek(this.state.currentTime + distance);
 		this.startDelayedMiniFeedbackHide();
-	}
+	};
 
 	/**
 	 * Changes the playback speed via [selectPlaybackRate()]{@link moonstone/VideoPlayer.VideoPlayer#selectPlaybackRate}.
@@ -1361,7 +1361,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.stopDelayedMiniFeedbackHide();
 		this.clearPulsedPlayback();
 		this.showFeedback();
-	}
+	};
 
 	/**
 	 * Changes the playback speed via [selectPlaybackRate()]{@link moonstone/VideoPlayer.VideoPlayer#selectPlaybackRate}.
@@ -1425,7 +1425,7 @@ const VideoPlayerBase = class extends React.Component {
 		this.stopDelayedMiniFeedbackHide();
 		this.clearPulsedPlayback();
 		this.showFeedback();
-	}
+	};
 
 	// Creates a proxy to the video node if Proxy is supported
 	videoProxy = typeof Proxy !== 'function' ? null : new Proxy({}, {
@@ -1441,7 +1441,7 @@ const VideoPlayerBase = class extends React.Component {
 		set: (target, name, value) => {
 			return (this.video[name] = value);
 		}
-	})
+	});
 
 	/**
 	 * Returns a proxy to the underlying `<video>` node currently used by the VideoPlayer
@@ -1452,11 +1452,11 @@ const VideoPlayerBase = class extends React.Component {
 	 */
 	getVideoNode = () => {
 		return this.videoProxy || this.video;
-	}
+	};
 
 	areControlsVisible = () => {
 		return this.state.mediaControlsVisible;
-	}
+	};
 
 	/**
 	 * Sets the playback rate type (from the keys of [playbackRateHash]{@link moonstone/VideoPlayer.VideoPlayer#playbackRateHash}).
@@ -1466,7 +1466,7 @@ const VideoPlayerBase = class extends React.Component {
 	 */
 	selectPlaybackRates = (cmd) => {
 		this.playbackRates = this.props.playbackRateHash[cmd];
-	}
+	};
 
 	/**
 	 * Changes [playbackRate]{@link moonstone/VideoPlayer.VideoPlayer#playbackRate} to a valid value
@@ -1481,7 +1481,7 @@ const VideoPlayerBase = class extends React.Component {
 		}
 
 		return idx % this.playbackRates.length;
-	}
+	};
 
 	/**
 	 * Retrieves the playback rate name.
@@ -1492,7 +1492,7 @@ const VideoPlayerBase = class extends React.Component {
 	 */
 	selectPlaybackRate = (idx) => {
 		return this.playbackRates[idx];
-	}
+	};
 
 	/**
 	 * Sets [playbackRate]{@link moonstone/VideoPlayer.VideoPlayer#playbackRate}.
@@ -1520,7 +1520,7 @@ const VideoPlayerBase = class extends React.Component {
 			// Set native playback rate
 			this.video.playbackRate = pbNumber;
 		}
-	}
+	};
 
 	/**
 	 * Calculates the time that has elapsed since. This is necessary for browsers until negative
@@ -1538,9 +1538,9 @@ const VideoPlayerBase = class extends React.Component {
 		this.stopDelayedMiniFeedbackHide();
 		this.clearPulsedPlayback();
 		this.startRewindJob();	// Issue another rewind tick
-	}
+	};
 
-	rewindJob = new Job(this.rewindManually, 100)
+	rewindJob = new Job(this.rewindManually, 100);
 
 	/**
 	 * Starts rewind job.
@@ -1550,7 +1550,7 @@ const VideoPlayerBase = class extends React.Component {
 	startRewindJob = () => {
 		this.rewindBeginTime = perfNow();
 		this.rewindJob.start();
-	}
+	};
 
 	/**
 	 * Stops rewind job.
@@ -1559,7 +1559,7 @@ const VideoPlayerBase = class extends React.Component {
 	 */
 	stopRewindJob = () => {
 		this.rewindJob.stop();
-	}
+	};
 
 	/**
 	 * Implements custom rewind functionality (until browsers support negative playback rate).
@@ -1569,7 +1569,7 @@ const VideoPlayerBase = class extends React.Component {
 	beginRewind = () => {
 		this.send('pause');
 		this.startRewindJob();
-	}
+	};
 
 	//
 	// Handled Media events
@@ -1582,12 +1582,12 @@ const VideoPlayerBase = class extends React.Component {
 			// Specific state variables are included in the outgoing calback payload, not all of them
 			...this.getMediaState()
 		};
-	}
+	};
 
 	disablePointerMode = () => {
 		Spotlight.setPointerMode(false);
 		return true;
-	}
+	};
 
 	handleKeyDownFromControls = this.handle(
 		// onKeyDown is used as a proxy for when the title has been read because it can only occur
@@ -1596,14 +1596,14 @@ const VideoPlayerBase = class extends React.Component {
 		forKey('down'),
 		this.disablePointerMode,
 		this.hideControls
-	)
+	);
 
 	//
 	// Player Interaction events
 	//
 	onVideoClick = () => {
 		this.toggleControls();
-	}
+	};
 
 	onSliderChange = ({value}) => {
 		const time = value * this.state.duration;
@@ -1612,9 +1612,9 @@ const VideoPlayerBase = class extends React.Component {
 
 		this.seek(time);
 		this.sliderScrubbing = false;
-	}
+	};
 
-	sliderTooltipTimeJob = new Job((time) => this.setState({sliderTooltipTime: time}), 20)
+	sliderTooltipTimeJob = new Job((time) => this.setState({sliderTooltipTime: time}), 20);
 
 	handleKnobMove = (ev) => {
 		this.sliderScrubbing = true;
@@ -1634,7 +1634,7 @@ const VideoPlayerBase = class extends React.Component {
 				this.announce(`${$L('jump to')} ${knobTime}`);
 			}
 		}
-	}
+	};
 
 	handleSliderFocus = () => {
 		const seconds = Math.floor(this.sliderKnobProportion * this.video.duration);
@@ -1658,7 +1658,7 @@ const VideoPlayerBase = class extends React.Component {
 
 			this.announce(`${$L('jump to')} ${knobTime}`);
 		}
-	}
+	};
 
 	handleSliderBlur = () => {
 		this.sliderScrubbing = false;
@@ -1668,7 +1668,7 @@ const VideoPlayerBase = class extends React.Component {
 			feedbackVisible: true,
 			sliderTooltipTime: currentTime
 		}));
-	}
+	};
 
 	slider5WayPressJob = new Job(() => {
 		this.setState({slider5WayPressed: false});
@@ -1699,16 +1699,16 @@ const VideoPlayerBase = class extends React.Component {
 		} else {
 			this.activityDetected();
 		}
-	}
+	};
 
 	onJumpBackward = this.handle(
 		forwardJumpBackward,
 		() => this.jump(-1 * this.props.jumpBy)
-	)
+	);
 	onJumpForward = this.handle(
 		forwardJumpForward,
 		() => this.jump(this.props.jumpBy)
-	)
+	);
 
 	handleToggleMore = ({showMoreComponents}) => {
 		if (!showMoreComponents) {
@@ -1725,31 +1725,31 @@ const VideoPlayerBase = class extends React.Component {
 			titleVisible: true,
 			announce: announce < AnnounceState.INFO ? AnnounceState.INFO : AnnounceState.DONE
 		}));
-	}
+	};
 
 	handleMediaControlsClose = (ev) => {
 		this.hideControls();
 		ev.stopPropagation();
-	}
+	};
 
 	setPlayerRef = (node) => {
 		// TODO: We've moved SpotlightContainerDecorator up to allow VP to be spottable but also
 		// need a ref to the root node to query for children and set CSS variables.
 		// eslint-disable-next-line react/no-find-dom-node
 		this.player = ReactDOM.findDOMNode(node);
-	}
+	};
 
 	setVideoRef = (video) => {
 		this.video = video;
-	}
+	};
 
 	setTitleRef = (node) => {
 		this.titleRef = node;
-	}
+	};
 
 	setAnnounceRef = (node) => {
 		this.announceRef = node;
-	}
+	};
 
 	getControlsAriaProps () {
 		if (this.state.announce === AnnounceState.TITLE) {